wip(cli): check_arb refactor
